pyproject.toml
[project]
name = "papers-2-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bson>=0.5.10",
    "dateutils>=0.6.12",
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "flask>=3.1.0",
    "flask-core>=2.9.0",
    "flask-cors>=5.0.1",
    "flask-limiter>=3.12",
    "flask-pymongo>=3.0.1",
    "flask-talisman>=1.1.0",
    "flask-wtf>=1.2.2",
    "google>=3.0.0",
    "google-genai>=1.20.0",
    "gradio>=5.32.1",
    "httpx>=0.28.1",
    "jwt>=1.3.1",
    "openai>=1.84.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.9.1",
    "pydantic[email]>=2.11.4",
    "pymongo>=4.12.1",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.4.0",
    "requests>=2.32.3",
    "schedule>=1.2.2",
    "slowapi>=0.1.9",
    "uvicorn[standard]>=0.34.2",
    "waitress>=3.0.2",
]

[tool.ruff]
target-version = "py312"
line-length = 120
